Fix freeze on load caused by helm issue.
There seems to be an issue between helm and the specific version
of emacs used by Arch. This commit adds a patch that should resolve the
issue.

When Emacs 25 is available we should remove the patch and see if the issue
still exists.

See: https://github.com/emacs-helm/helm/issues/1000