Make the "/code" mounted folder more permissive in Vagrant
We want to make sure the user has full access to the mounted
code folder. Unfortunately we can't just change it's owner
using the vagrant settings as the user might not exist if it's
the first provision.

Instead we just give everyone read/write/execute access which
they'll have already if they're mounting it from windows. If
this turns out to be a problem we can manually mount the folder
with a script.