Ensure `.nix-defexpr` changes stick around after restart
It looks like the `/etc/profile` script included with nix that sets up
`.nix-defexpr` will overwrite `.nix-defexpr` if it is non-existent or
*if it is a symlink*!

We don't want it to do that and I didn't want to modify nixos directly nor do
I want to run the install script in `~/.profile` if I can avoid it.

Instead we can ensure `~/.nix-defexpr` is a real directory which should not
be overwriten, then we just need to clear out all the channel symlinks and
replace them a symlink to our chosen nixpkgs. I've opted to only use one
nixpkgs in `~/.nix-defexpr` since most `nix-env` commands seem to traverse
both channels and they're already slow enough.

Hopefully one day `nix-env` will support `NIX_PATH` or be replaced with
something that does. If not, this will just have to do.